+++ src/prepareCommandOptions.js
@@ -1,0 +1,12 @@
+/**
+ * Prepares the commandOptions string to be valid in case that is not included in the params
+ * @param {string} commandOptions
+ * @returns {*}
+ */
+ module.exports = function prepareCommandOptions(commandOptions) {
+    if (commandOptions == null) {
+        return '';
+    }
+
+    return commandOptions;
+};
